feature: Transfere o aplicativo para o método de visualização Stack Navigator.
